StarWarsBot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asyncio==4.0.0
attrs==25.4.0
black==25.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
propcache==0.4.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
ruff==0.14.8
typing_extensions==4.15.0
urllib3==2.6.0
yarl==1.22.0
